requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
openai==1.59.6
pydantic==2.10.4
pydantic-settings==2.7.1
aiohttp==3.11.11
websockets==14.1
cryptography>=44.0.0
numpy==2.2.1
soundfile==0.12.1
python-multipart==0.0.20
jinja2==3.1.4

# ASR
faster-whisper>=1.0.0

# TTS
edge-tts>=6.1.0

# VAD - 使用 Silero 模型
torch>=2.0.0
torchaudio>=2.0.0
